Added status update of all presentations after deleting all files

+++ lib/features/presentation_view/screens/loading_screen/loading_screen_vm.dart
@@ -10,8 +10,12 @@ class LoadingScreenVm extends ChangeNotifier {
     required DeletePresentation deletePresentation,
   }) : _deletePresentation = deletePresentation;
 
-  Future<void> init(BuildContext context) async {
+  Future<void> init(
+    BuildContext context,
+    VoidCallback refreshPresentationsStates,
+  ) async {
     _context = context;
+    _refreshPresentationsStates = refreshPresentationsStates;
     presentations = Mocks().presentationsMock;
   }
 
@@ -19,6 +23,7 @@ class LoadingScreenVm extends ChangeNotifier {
   ///
 
   late final BuildContext _context;
+  late final VoidCallback _refreshPresentationsStates;
   late final List<PresentationEntity> presentations;
 
   /// Reactive properties
@@ -39,9 +44,30 @@ class LoadingScreenVm extends ChangeNotifier {
   ///
 
   Future<void> deleteAll() async {
+    final userConfirm = await showDialog<bool>(
+      context: _context,
+      builder: (context) => AlertDialog(
+        title: const Text('Подтверждение'),
+        content: const Text('Удалить все загруженные файлы презентаций?'),
+        actions: [
+          OutlinedButton(
+            onPressed: () => Navigator.of(_context).pop(false),
+            child: const Text('Отмена'),
+          ),
+          OutlinedButton(
+            onPressed: () => Navigator.of(_context).pop(true),
+            child: const Text('Удалить всё'),
+          ),
+        ],
+      ),
+    );
+    if (userConfirm != true) {
+      return;
+    }
     loading = true;
     try {
       await _deletePresentation.deleteAll();
+      _refreshPresentationsStates();
       _showToast('Удаление завершено');
     } catch (error) {
       _showToast('Ошибка удаления: $error');
